Zp3 slotter

+++ src/App.jsx
@@ -52,6 +52,7 @@ import {
   jobMetrics,
   zitaMatchScreen,
   calendarRoute,
+  meetingScheduler,
 } from './appRoutesPath';
 import CreateJdWithNonDs from './modules/createjdmodule/CreateJdWithNonDs';
 import MyJobPostingScreen from './modules/myjobposting/MyJobPostingScreen';
@@ -78,6 +79,9 @@ import SourcingPerformance from './modules/reportsmodule/SourcingPerformance';
 import DashBoardScreen from './modules/dashboardmodule/empdashboard/DashBoardScreen';
 import CheckSignUpActivate from './modules/SignUp/CheckSignUpActivate';
 
+import MeetingSchedulerScreen from './modules/meetingscheduler/meetingSchedulerScreen';
+import Slotter1 from './modules/meetingscheduler/events/Slotter1';
+import PreviewTabs from './modules/meetingscheduler/events/PreviewTab';
 
 axios.defaults.baseURL = process.env.REACT_APP_API_ENDPOINT;
 
@@ -92,7 +96,7 @@ const App = () => {
       return response;
     },
     (error) => {
-      if (error.response.status === 401) {
+      if (error?.response && error?.response?.status === 401) {
         window.location.replace(`${window.location.origin + '/login'}`);
         localStorage.clear();
         sessionStorage.clear();
@@ -100,7 +104,6 @@ const App = () => {
       return error;
     },
   );
-
 
   // useEffect(() => {
   //   const query = parse(location.search);
@@ -339,12 +342,20 @@ const App = () => {
       exact: true,
       noPermission: true,
       isNav: true,
+      isside:true 
+    },
+    {
+      path: meetingScheduler,
+      component: MeetingSchedulerScreen,
+      exact: true,
+      noPermission: true,
+      isNav: true,
       isside:true
     },
+     
 
     { path: '/logout', component: Logout, exact: true, noPermission: true },
   ];
-
 
   // Candidate Route
 
@@ -396,34 +407,35 @@ const App = () => {
         <Route path={'/check_activate'} component={CheckSignUpActivate} />
         {/* <Route path={''} component={NotFound} /> */}
         {/* <Route path="/calendar" component={Calendar} /> */}
+        <Route path="/slotter" component={Slotter1}/>
+        <Route path="/event_preview" component={PreviewTabs}/>
         {localStorage.getItem('loginUserCheck') === 'true' ||
-          localStorage.getItem('loginUserCheck') === null
+        localStorage.getItem('loginUserCheck') === null
           ? permissionRoutes.map(
-            (route) =>
-              route.noPermission && (
-                <ProtectedRoute
-                  key={route.path}
-                  exact={route.exact}
-                  path={route.path}
-                  component={route.component}
-                  notIsNav={route.isNav}
+              (route) =>
+                route.noPermission && (
+                  <ProtectedRoute
+                    key={route.path}
+                    exact={route.exact}
+                    path={route.path}
+                    component={route.component}
+                    notIsNav={route.isNav}
                   isside={route.isside}
-                />
-              ),
-          )
+                  />
+                ),
+            )
           : candidateRoutes.map((route) => (
-            <ProtectedRouteCandidate
-              key={route.path}
-              exact={route.exact}
-              path={route.path}
-              component={route.component}
-              notIsNav={route.isNav}
-            />
-          ))}
+              <ProtectedRouteCandidate
+                key={route.path}
+                exact={route.exact}
+                path={route.path}
+                component={route.component}
+                notIsNav={route.isNav}
+              />
+            ))}
       </Switch>
     </BrowserRouter>
   );
 };
 
 export default App;
-
